Sdk dev

See merge request wd_intelligent_building/lyzn-backend/huiyun4.0/service_edge/hdwa-project-sdk!9

+++ src/main/java/com/hdwa/sdk/utils/DataUtil.java
@@ -21,7 +21,7 @@ public class DataUtil {
         } else if (value instanceof Float) {
             return ((Float) value).doubleValue();
         } else if (value instanceof Double) {
-            return ((Double) value).doubleValue();
+            return value;
         } else if (value instanceof BigDecimal) {
             return ((BigDecimal) value).doubleValue();
         } else {
